fixing kiva's setup.py to build under linux, fixing some build warnings in our bundled agg 2.4 and our numeric.i swig wrapper

+++ enthought/kiva/agg/agg-24/include/agg_image_accessors.h
@@ -64,8 +64,9 @@ namespace agg
         {
             m_x = m_x0 = x;
             m_y = y;
+            // comparison between signed and unsigned
             if(y >= 0 && y < (int)m_pixf->height() &&
-               x >= 0 && x+len <= (int)m_pixf->width())
+               x >= 0 && (int)(x+len) <= (int)m_pixf->width())
             {
                 return m_pix_ptr = m_pixf->pix_ptr(x, y);
             }
